+++ lib/Db/FileElementMapper.php
@@ -1,0 +1,27 @@
+<?php
+
+namespace OCA\Libresign\Db;
+
+use OCP\AppFramework\Db\QBMapper;
+use OCP\IDBConnection;
+
+/**
+ * Class FileElementsMapper
+ *
+ * @package OCA\Libresign\DB
+ *
+ * @codeCoverageIgnore
+ * @method FileUser insert(FileUser $entity)
+ * @method FileUser update(FileUser $entity)
+ * @method FileUser insertOrUpdate(FileUser $entity)
+ * @method FileUser delete(FileUser $entity)
+ */
+class FileElementMapper extends QBMapper {
+
+	/**
+	 * @param IDBConnection $db
+	 */
+	public function __construct(IDBConnection $db) {
+		parent::__construct($db, 'libresign_file_element');
+	}
+}